refactor: receiver 관련 도메인 모델 패키지 이동 및 presentation 모듈 추가
- `feature:receiver:presentation` 모듈 신규 생성 및 Compose/Hilt 등 기초 빌드 설정 추가
- `SenderMessageInfo`, `ReceivedAccountCredentials`, `ReceivedPlaylistDetail`, `ReceivedPlaylistSong` 등 수신자 관련 도메인 모델을 `com.afternote.feature.receiver.domain` 패키지로 이동
- `ReceiverAuthRepository` 인터페이스의 패키지 위치 변경 반영
- 도메인 모델 및 리포지토리 위치 변경에 따른 `afternote` 모듈 내 데이터 매퍼 및 리포지토리 구현체의 임포트 경로 수정
