+++ feature/receiver/presentation/build.gradle.kts
@@ -1,0 +1,22 @@
+plugins {
+    id("afternote.android.library.compose")
+    id("afternote.android.hilt")
+}
+
+android {
+    namespace = "com.afternote.feature.receiver.presentation"
+}
+
+dependencies {
+    implementation(projects.feature.receiver.domain)
+    implementation(projects.core.domain)
+    implementation(projects.core.model)
+    implementation(projects.core.common)
+    implementation(projects.core.ui)
+    implementation(libs.androidx.activity.compose)
+    implementation(libs.android.navigation.compose)
+    implementation(libs.androidx.hilt.navigation.compose)
+    implementation(libs.hilt.navigation)
+    implementation(libs.androidx.lifecycle.viewmodel.ktx)
+    implementation(libs.androidx.lifecycle.runtime.compose)
+}
